fixed invalid types of arguments in ElectionsControl::getState()

+++ app/Components/ElectionsControl.php
@@ -40,12 +40,12 @@ class ElectionsControl extends \Nette\Application\UI\Control {
    */
   protected function getState(): string {
     $fakeDay = new \DateTime;
-    $fakeDay->setDate(date("Y"), date("n"), date("j"));
+    $fakeDay->setDate((int) date("Y"), (int) date("n"), (int) date("j"));
     if($fakeDay->format("j") <= 7) {
       $state = "results";
     } else {
       $date = new \DateTime;
-      $date->setDate(date("Y"), date("n") + 1, 1);
+      $date->setDate((int) date("Y"), (int) date("n") + 1, 1);
       $date->setTime(0, 0, 0);
       $date->modify("-7 days");
       if($fakeDay->getTimestamp() > $date->getTimestamp()) $state = "voting";
